# src/main/resources/static/swagger-config.yml
openapi: 3.0.0
info:
  description: "Box Demo application service"
  version: "1.0.0"
  title: "Box Demo"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "blue_farid@hotmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

servers:
  - url: http://localhost:8080/api
tags:
  - name: "requests"
    description: "requests descriptions"
paths:
  /order:
    post:
      tags:
        - "requests"
      summary: "Add new BoxOrder to the resource"
      operationId: "postOrder"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoxOrder'
            examples:
              BoxOrder:
                value:
                  ownerFullName: "Farid Masjedi"
                  ownerPhoneNumber: "09123456789"
                  sourceFullName: "Farid Masjedi"
                  sourcePhoneNumber: "09123456789"
                  sourceAddressBase: "Tehran, Qeytarieh, Andarzgoo"
                  sourceAddressHouseNumber: "4"
                  sourceAddressHomeUnit: "3"
                  sourceComment: "Sample description for the order"
                  destinations:
                    destination_1:
                      fullName: "Neda Masjedi"
                      phoneNumber: "09123456788"
                      addressBase: "Tehran, Sattarkhan"
                      addressHouseNumber: "10"
                      addressHomeUnit: "3"
                      comment: "some comment for this destination"
                      priceRange: "UP_TO_ONE"
                  orderType: "BIKE"
      responses:
        200:
          description: order has been saved successfully!
        400:
          description: duplicate order with same id!
        500:
          description: Internal Server Error!
    put:
      tags:
        - "requests"
      summary: "update the box order with the specific id"
      operationId: "putOrder"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoxOrder'
            examples:
              BoxOrder:
                value:
                  id: 1
                  ownerFullName: "Farid Masjedi"
                  ownerPhoneNumber: "09123456789"
                  sourceFullName: "Farid Masjedi"
                  sourcePhoneNumber: "09123456789"
                  sourceAddressBase: "Tehran, Qeytarieh, Andarzgoo"
                  sourceAddressHouseNumber: "4"
                  sourceAddressHomeUnit: "3"
                  sourceComment: "Sample description for the order"
                  destinations:
                    destination_1:
                      fullName: "Neda Masjedi"
                      phoneNumber: "09123456788"
                      addressBase: "Tehran, Sattarkhan"
                      addressHouseNumber: "10"
                      addressHomeUnit: "3"
                      comment: "some comment for this destination"
                      priceRange: "UP_TO_ONE"
                  orderType: "BIKE"
      responses:
        200:
          description: order has been updated successfully!
        404:
          description: not found!
        500:
          description: Internal Server Error!


  /order/{orderId}:
    get:
      tags:
        - "requests"
      summary: "returns the BoxOrder"
      operationId: "getBoxOrder"
      parameters:
        - in: path
          required: true
          name: orderId
          description: order id
      responses:
        200:
          description: "order has been returned!"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoxOrder'
              examples:
                BoxOrder:
                  value:
                    ownerFullName: "Farid Masjedi"
                    ownerPhoneNumber: "09123456789"
                    sourceFullName: "Farid Masjedi"
                    sourcePhoneNumber: "09123456789"
                    sourceAddressBase: "Tehran, Qeytarieh, Andarzgoo"
                    sourceAddressHouseNumber: "4"
                    sourceAddressHomeUnit: "3"
                    sourceComment: "Sample description for the order"
                    destinations:
                      destination_1:
                       fullName: "Neda Masjedi"
                       phoneNumber: "09123456788"
                       addressBase: "Tehran, Sattarkhan"
                       addressHouseNumber: "10"
                       addressHomeUnit: "3"
                       comment: "some comment for this destination"
                       priceRange: "UP_TO_ONE"
                    orderType: "BIKE"
        404:
          description: "order not found!"
        500:
          description: "Internal server error!"
    delete:
      tags:
        - "requests"
      summary: "deletes the BoxOrder"
      operationId: "removeBoxOrder"
      parameters:
        - in: path
          required: true
          name: orderId
          description: order id
      responses:
        200:
          description: "order has been removed!"
        404:
          description: "order not found!"
        500:
          description: "Internal server error!"


components:
  schemas:
    BoxOrder:
      type: object
      properties:
        id:
          type: number
          required: false
          description: order id
        ownerFullName:
          type: string
          required: true
          description: client full name
        ownerPhoneNumber:
          type: string
          required: true
          description: client phone number
        sourceFullName:
          type: string
          required: true
          description: source full name
        sourcePhoneNumber:
          type: number
          required: true
          description: source phone number
        sourceAddressBase:
          type: string
          required: true
          description: source base address
        sourceAddressHouseNumber:
          type: integer
          description: source house number value
          required: true
        sourceAddressHomeUnit:
          type: integer
          description: source home unit value
          required: true
        sourceComment:
          type: string
          description: source comments
        destinations:
          type: array
          description: list of destination objects - at least one element.
          required: true
        orderType:
          type: string
          description: order type - MUST be 'bike' or 'car'.